Shift more things from RWL.cc to RWLInternal.h
